here are a bunch of cleanups and refactors to set the stage for breaking cell scanning out of the wifireceiver
