+++ wiglewifiwardriving/src/main/java/net/wigle/wigleandroid/MappingFragment.java
@@ -63,7 +63,7 @@ import net.wigle.wigleandroid.background.QueryThread;
 import net.wigle.wigleandroid.db.DatabaseHelper;
 import net.wigle.wigleandroid.model.ConcurrentLinkedHashMap;
 import net.wigle.wigleandroid.model.Network;
-import net.wigle.wigleandroid.util.WiGLEToast;
+import net.wigle.wigleandroid.ui.WiGLEToast;
 
 /**
  * show a map!
