+++ wiglewifiwardriving/src/main/java/net/wigle/wigleandroid/ui/WiGLEToast.java
@@ -1,4 +1,4 @@
-package net.wigle.wigleandroid.util;
+package net.wigle.wigleandroid.ui;
 
 import android.app.Activity;
 import android.os.Build;
